+++ CHANGELOG.md
@@ -9,6 +9,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [0.7.1] - 2024-03-05
+
+### Added
+
+- Implemented `From<Configuration>` for `dogstatsd::Options`
+
+### Changed
+
+- `Config` is now `Configuration`
+
+### Removed
+
+- The `Configuration` trait
+- `Country` and `Environment` serialization
+
+---
+
 ## [0.7.0] - 2024-03-04
 
 ### Changed
@@ -111,8 +128,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.5]
 
-
-[Unreleased]: https://github.com/primait/prima_datadog.rs/compare/0.7.0...HEAD
+[Unreleased]: https://github.com/primait/prima_datadog.rs/compare/0.7.1...HEAD
+[0.7.1]: https://github.com/primait/prima_datadog.rs/compare/0.7.0...0.7.1
 [0.7.0]: https://github.com/primait/prima_datadog.rs/compare/0.7.0...0.7.0
 [0.6.0]: https://github.com/primait/prima_datadog.rs/compare/0.5.0...0.6.0
 [0.5.0]: https://github.com/primait/prima_datadog.rs/compare/0.4.0...0.5.0